added details for groups
