+++ src/components/attribute-form/attribute-form.css
@@ -1,0 +1,27 @@
+.kc-attributes__table {
+  /* even though the table is borderless, make the border under the th transparent */
+  --pf-c-table--border-width--base: 0;
+  --pf-c-table--m-compact--cell--first-last-child--PaddingLeft: 0;
+}
+
+.kc-attributes__plus-icon {
+  /* shift the button left to adjust for table cell padding */
+  margin-left: calc(var(--pf-global--spacer--md) * -1);
+}
+
+.pf-c-button.kc-attributes__minus-icon {
+  /* shift the button left to adjust for table cell padding */
+  margin-left: calc(var(--pf-global--spacer--md) * -1);
+  color: var(--pf-c-button--m-plain--Color);
+}
+
+.kc-add-button {
+  margin-left: var(--pf-global--spacer--lg);
+}
+
+.kc-attributes__action-group {
+  /* subtract the padding at the bottom of the table from the action group margin */
+  --pf-c-form__group--m-action--MarginTop: calc(
+    var(--pf-global--spacer--2xl) - var(--pf-global--spacer--sm)
+  );
+}
