+++ src/realm-roles/RealmRolesSection.css
@@ -1,30 +1,4 @@
-.kc-role-attributes__table {
-  /* even though the table is borderless, make the border under the th transparent */
-  --pf-c-table--border-width--base: 0;
-  --pf-c-table--m-compact--cell--first-last-child--PaddingLeft: 0;
-}
 
-.kc-role-attributes__plus-icon {
-  /* shift the button left to adjust for table cell padding */
-  margin-left: calc(var(--pf-global--spacer--md) * -1);
-}
-
-.pf-c-button.kc-role-attributes__minus-icon {
-  /* shift the button left to adjust for table cell padding */
-  margin-left: calc(var(--pf-global--spacer--md) * -1);
-  color: var(--pf-c-button--m-plain--Color);
-}
-
-.kc-add-role-button {
-  margin-left: var(--pf-global--spacer--lg);
-}
-
-.kc-role-attributes__action-group {
-  /* subtract the padding at the bottom of the table from the action group margin */
-  --pf-c-form__group--m-action--MarginTop: calc(
-    var(--pf-global--spacer--2xl) - var(--pf-global--spacer--sm)
-  );
-}
 
 .kc-who-will-appear-button {
   padding-left: 0px;
